Core data viewmodel
